+++ .replit
@@ -1,0 +1,1 @@
+run="./repl.deploy node index.js"
