+++ package.json
@@ -9,6 +9,7 @@
     "openai": "^2.0.5",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
     "sass": "^1.51.0",
     "web-vitals": "^2.1.4"
