+++ src/Components/Home/Home.jsx
@@ -1,0 +1,5 @@
+const Home = () => {
+  return <div className='Home'>home container</div>;
+};
+
+export default Home;
