+++ src/Styles/_App.scss
@@ -1,4 +1,4 @@
-.App {
+.AIContent {
   display: flex;
   flex-direction: column;
   justify-content: center;
